Refactor common domain and llm config to specialized crates
- Moved `tardis_common::domain` to `tardis_gallifrey::domain`.
- Moved `tardis_common::llm` to `tardis_vortex::config`.
- Deleted `common/src/domain.rs` and `common/src/llm.rs`.
- Updated all import paths in `chronos`, `gallifrey`, `vortex`.
- Fixed clippy warnings in `chronos` and `telemetry`.
- Enforced Razor's philosophy of de-abstraction and flattening.
