+++ chronos/src/experimental/dreamer.rs
@@ -6,7 +6,7 @@ use crate::error::ChronosResult;
 use async_trait::async_trait;
 use std::collections::HashMap;
 use std::sync::Arc;
-use tardis_common::domain::{Entity, Message};
+use tardis_gallifrey::domain::{Entity, Message};
 use tardis_common::id::EntityId;
 use tardis_gallifrey::Gallifrey;
 use tracing::{info, instrument};
@@ -126,7 +126,7 @@ impl DreamCatcher for MockDreamCatcher {
 mod tests {
     use super::*;
     use std::sync::Arc;
-    use tardis_common::domain::{Message, Role};
+    use tardis_gallifrey::domain::{Message, Role};
     use tardis_common::id::EntityId;
 
     #[tokio::test]
