+++ chronos/src/experimental/prophecy.rs
@@ -9,9 +9,9 @@ use crate::error::{ChronosError, ChronosResult};
 use crate::experimental::psychic_paper::{Intent, PsychicPaper};
 use std::sync::Arc;
 use std::time::Duration;
-use tardis_common::domain::Entity;
+use tardis_gallifrey::domain::Entity;
 use tardis_common::id::{EntityId, ModelHandle};
-use tardis_common::llm::InferenceParams;
+use tardis_vortex::config::InferenceParams;
 use tardis_common::temporal::{BiTemporalInterval, TimeRange};
 use tardis_vortex::Vortex;
 use tracing::{info, instrument};
