+++ chronos/src/pipeline/analyzer.rs
@@ -263,6 +263,7 @@ impl Default for QueryAnalyzer {
 
 #[cfg(test)]
 #[allow(clippy::unwrap_used)]
+#[allow(clippy::panic)]
 mod tests {
     use super::*;
