+++ gallifrey/src/stores/conversation.rs
@@ -9,7 +9,7 @@ use crate::error::{GallifreyError, GallifreyResult};
 use chrono::Utc;
 use std::collections::HashMap;
 use std::sync::RwLock;
-pub use tardis_common::domain::{Message, Role, Session};
+pub use crate::domain::{Message, Role, Session};
 use tardis_common::{EntityId, SessionId};
 
 /// The conversation store.
