+++ gallifrey/src/stores/system_state.rs
@@ -9,7 +9,7 @@ use crate::error::{GallifreyError, GallifreyResult};
 use chrono::{DateTime, Utc};
 use std::collections::HashMap;
 use std::sync::RwLock;
-pub use tardis_common::domain::{Change, Snapshot, SnapshotTrigger, SystemState};
+pub use crate::domain::{Change, Snapshot, SnapshotTrigger, SystemState};
 use tardis_common::SnapshotId;
 
 /// The system state store.
