+++ vortex/src/lib.rs
@@ -38,6 +38,7 @@
 #![warn(clippy::pedantic)]
 #![allow(clippy::module_name_repetitions)]
 
+pub mod config;
 pub mod error;
 pub mod inference;
 pub mod loader;
@@ -45,9 +46,9 @@ pub mod model;
 pub mod tokenizer;
 
 // Re-export main types
+pub use config::{InferenceParams, ModelLoadConfig};
 pub use error::{VortexError, VortexResult};
 pub use inference::Vortex;
 pub use loader::ModelPreset;
 pub use model::{ModelHandle, ModelInfo, ModelRegistry};
-pub use tardis_common::llm::{InferenceParams, ModelLoadConfig};
 pub use tokenizer::TokenizerService;
